+++ opentelemetry-sdk/src/opentelemetry/sdk/_logs/_internal/__init__.py
@@ -170,7 +170,7 @@ class LogRecord(APILogRecord):
     pertinent to the event being logged.
     """
 
-    def __init__(
+    def __init__(  # pylint: disable=too-many-positional-arguments
         self,
         timestamp: Optional[int] = None,
         observed_timestamp: Optional[int] = None,